Update "Run Katago with Core ML in Apple silicon"
This commit introduces a guide to convert a KataGo network to a Core ML model using TensorFlow 1.5 and generating an Xcode project from KataGo. It also mentions the support for model version 11 trained on PyTorch, allowing for the conversion of a KataGo network of PyTorch to a Core ML model without needing an Intel-based Macbook. See the linked release page for more information about KataGo Core ML model.