Add post on optimizing KataGo performance with CMA-ES
In this commit, a new blog post is added that details the process and results of using the Covariance Matrix Adaptation Evolution Strategy (CMA-ES) for optimizing the parameters of KataGo, a Go-playing AI. The post covers the game configurations used, the parameter search ranges, the CMA-ES configurations, and the evaluation of the CMA-ES solutions. It also provides a comparison of the default and CMA-ES KataGo parameters, as well as a strength difference analysis. Finally, a link to the source code for the experiment is provided for those interested in further exploration.